[MASTERDIET-12] Calculations of user plan

+++ src/main/java/pl/agh/edu/master_diet/core/model/database/Activity.java
@@ -1,17 +1,8 @@
 package pl.agh.edu.master_diet.core.model.database;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "activity")
@@ -32,11 +23,11 @@ public class Activity {
     private String description;
 
     @Column(nullable = false)
-    private Float burnedCalories;
+    private Double burnedCalories;
 
     private Integer defaultReps;
 
     private Integer defaultSets;
 
-    private Float defaultTime;
+    private Double defaultTime;
 }
