+++ src/main/java/pl/agh/edu/master_diet/core/model/database/UserPlan.java
@@ -1,24 +1,11 @@
 package pl.agh.edu.master_diet.core.model.database;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import pl.agh.edu.master_diet.core.model.rest.userPlan.UserPlanResponse;
 import pl.agh.edu.master_diet.core.model.shared.ActivityLevel;
 import pl.agh.edu.master_diet.core.model.shared.Goal;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MapsId;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "user_plan")
@@ -37,13 +24,13 @@ public class UserPlan {
     private Integer calories;
 
     @Column(nullable = false)
-    private Float proteins;
+    private Integer proteins;
 
     @Column(nullable = false)
-    private Float fat;
+    private Integer fat;
 
     @Column(nullable = false)
-    private Float carbohydrates;
+    private Integer carbohydrates;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
@@ -54,13 +41,13 @@ public class UserPlan {
     private Goal goal;
 
     @Column(nullable = false)
-    private Float currentWeight;
+    private Double currentWeight;
 
     @OneToOne
     @MapsId
     private User user;
 
-    public UserPlanResponse map2UserPlanResponse(){
+    public UserPlanResponse map2UserPlanResponse() {
         return UserPlanResponse.builder()
                 .userId(getId())
                 .calories(getCalories())
